+++ eras/conway/impl/CHANGELOG.md
@@ -18,6 +18,8 @@
 * Change the Conway txInfo to allow Plutus V3
   NOTE - unlike V1 and V2, the ledger will no longer place the "zero ada" value
   in the script context for the transaction minting field.
+* Added instances for ConwayDelegsPredFailure:
+  `NoThunks`, `EncCBOR`, `DecCBOR`, and `Arbitrary`
 
 ## 1.1.0.0